+++ main/src/main/java/net/oneandone/stool/cli/Start.java
@@ -307,6 +307,7 @@ public class Start extends StageCommand {
 
     private String catalinaOpts(Ports ports, Stage stage) {
         List<String> opts;
+        String tomcatOpts;
 
         opts = new ArrayList<>();
 
@@ -318,6 +319,9 @@ public class Start extends StageCommand {
         } catch (ReadLinkException e) {
             throw new IllegalStateException(e);
         }
+
+        tomcatOpts = stage.macros().replace(stage.config().tomcatOpts);
+        opts.addAll(Separator.SPACE.split(tomcatOpts));
 
         opts.add("-Xmx" + stage.config().tomcatHeap + "m");